Fix SEGV when parsed file size is a whole number of pages.
The zero byte would go to a page on its own, not backed by the file, and
even though we CoW it immediately, cause a crash.

Instead of a smarter hack such as trying MAP_FIXED afterwards, just go
through the "non-seekable file" slow path; mmap was just a speed
optimization.

Note that zero bytes are illegal inside the 822 format.